% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var_cols.R
\name{add_var_cols}
\alias{add_var_cols}
\alias{add_var_cols2}
\title{Add Variance Columns for 3612 Reports}
\usage{
add_var_cols(D, vars = c("cost_pmpm", "claims_per_1000",
  "cost_per_claim"), reorder = T, format.pct = T)

add_var_cols2(D, vars = c("cost_pmpm", "claims_per_1000",
  "cost_per_claim"), reorder = T, format.pct = T,
  row_vars = c("product", "region", "mmcor_desc"))
}
\arguments{
\item{D}{A data.table}

\item{vars}{A character vector of column names to use to calculate new variance columns}

\item{reorder}{Logical, if \code{TRUE} (default) then reorders the columns in the order the vars are specified -
order goes as follows: prior, current, var, var_pct}

\item{format.pct}{Logical, if \code{TRUE} (default) then \code{pct_change()} will format the percentage variable}

\item{row_vars}{A character vector of column names that will serve as the unique rows of the final data.table -
usually product, region, and mmcor_desc}
}
\description{
These function adds variance and percent variance columns to a data.table for each of the vars specified.
     \code{add_var_cols2()} performs the same operations as \code{add_var_cols2()} but it also rehsapes the data.table
     using data.table's \code{dcast()} based on the variable names provided in \code{row_vars}.
}
\note{
\code{add_var_cols()} does not need to be reassigned to an R object, however \code{add_var_cols2()} does need to be reassigned.
     Variables cannot be included in both \code{var} and \code{row_vars} and variables appearing in neither are automatically dropped.
     If using \code{add_var_cols2()} a column \code{col} must exist containing two values 'prior' and 'current'.
}
