% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_all.r
\name{aggregate_all}
\alias{aggregate_all}
\title{Aggregate a data.frame and create 'ALL' values}
\usage{
aggregate_all(df, group_vars, agg_vars, fun = sum, no_all = NULL)
}
\arguments{
\item{df}{A data.table or data.frame}

\item{group_vars}{A character string of variable names to group by}

\item{agg_vars}{A character string of variable names to aggregate}

\item{fun}{The function with which to perform aggregation, default is \code{sum} (currently applies the function to all \code{agg_vars})}

\item{no_all}{A character string indicating which variables are not required to have an 'ALL' row in the final aggregated product, defaults to NULL}
}
\description{
This function aggregates the specified columns of a data.frame by the specified grouping variables,
It also prepares additional aggregations in which each grouping var is set to 'ALL'
}
\details{
If \code{no_all} == \code{group_vars} then this function merely performs a standard data.table aggregation,
     but if \code{no_all} != \code{group_vars} then this function returns all unique combinations of aggregations. Drops
     columns that do not appear in \code{group_vars} or \code{no_all}.
}
