% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards_ma.R
\name{backwards_ma}
\alias{backwards_ma}
\title{Backwards Looking Moving Average Function}
\usage{
backwards_ma(x, n)
}
\arguments{
\item{x}{Numeric object}

\item{n}{Number of periods}
}
\description{
Calculates the moving average of object x using the last n (including current) observations
}
\examples{
backwards_ma(x, n = 3)

}
