% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{comma.}
\alias{comma.}
\alias{comma.k}
\alias{comma.m}
\alias{comma.b}
\alias{comma.auto}
\title{Formatting Large Currency Values}
\usage{
comma.k(x, digits = 2L)

comma.m(x, digits = 2L)

comma.b(x, digits = 2L)

comma.auto(x, digits = 2L)
}
\arguments{
\item{x}{A numeric object}

\item{digits}{Number of rounded digits, default is 2}
}
\description{
These functions are extensions to formattable's comma() function.
\itemize{
   \item\code{.k} divides by 1,000 and adds "k" suffix;
   \item\code{.m} divides by 1,000,000 and adds "M" suffix;
   \item\code{.b} divides by 1,000,000,000 and adds "B" suffix.
   \item\code{.auto} automatically determines and applies whichever of the above formatting options is best.

}
}
\seealso{
[formattable::comma()]
}
