% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_3612.r
\name{convert_to_3612}
\alias{convert_to_3612}
\title{Prepares Data for 3-6-12 Reports}
\usage{
convert_to_3612(data)
}
\arguments{
\item{data}{A data.frame or data.table}
}
\description{
This function takes a data.table with the columns:
col {'prior', 'current'}
product
region
category
value
util
mm
If the columns product, region, or category are missing they are replaced with TRUE
Ensure there are no missing values, and the data has already been aggregated
}
