% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency.}
\alias{currency.}
\alias{currency.k}
\alias{currency.m}
\alias{currency.b}
\alias{currency.auto}
\title{Formatting Large Comma Values}
\usage{
currency.k(x, digits = 2L)

currency.m(x, digits = 2L)

currency.b(x, digits = 2L)

currency.auto(x, digits = 2L)
}
\arguments{
\item{x}{A numeric object}

\item{digits}{Number of rounded digits, default is 2}
}
\description{
These functions are extensions to formattable's currency() function.
\itemize{
   \item\code{.k} divides by 1,000 and adds "k" suffix;
   \item\code{.m} divides by 1,000,000 and adds "M" suffix;
   \item\code{.b} divides by 1,000,000,000 and adds "B" suffix.
   \item\code{.auto} automatically determines and applies whichever of the above formatting options is best.

}
}
\seealso{
[formattable::currency()]
}
