% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{n_top}
\alias{n_top}
\title{Return the Top n values of a data.frame}
\usage{
n_top(df, col, n, fun = function(x) {     -x }, keep_ordering = T)
}
\arguments{
\item{df}{A data.table or data.frame}

\item{col}{A character string of the column used to pick top values}

\item{n}{The number of top values to keep, nrow(df) < n then nrow(df) is used}

\item{fun}{The ordering function to apply to col. Default is function(x){-x} which keeps largest.}

\item{keep_ordering}{Logical, if TRUE then maintain original ordering, if FALSE order based on fun}
}
\description{
This function returns the top n values of a data.frame or data.table based on user provided function and column
}
\examples{
n_top(df, col = 'v1', n = 5) returns a data.table with the 5 largest values of v1 in the original ordering
n_top(df, col = 'v1', n = 10, fun = function(x){-abs(x)}, keep_ordering = F) returns a data.table with 10 largest absolute values of v1, ordered by descending absolute values of v1

}
