% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_with.r
\name{names_with}
\alias{names_with}
\title{Easily Select Elements from a Character Vector}
\usage{
names_with(names, with = NULL, without = NULL)
}
\arguments{
\item{names}{A character vector}

\item{with}{A character vector of items we want \code{names} to contain. Default to NULL.}

\item{without}{A character vector of items we do not want \code{names} to contain. Default to NULL.}
}
\description{
This function returns elements that contain characters specified in with and do not contain characters specified in without
}
\examples{

names <- c('one', 'two', 'three', 'three3')
with <- c('one', 'three')
without <- c('three3')

names_with(names, with, without) ##returns c('one', 'three')

names_with(names, with) ##returns c('one', 'three', 'three3')

}
