% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAto0.r
\name{NAto0}
\alias{NAto0}
\title{Convert NAs to Zero in data.table}
\usage{
NAto0(D, cols = names(D), value = 0)
}
\arguments{
\item{D}{A data.table or data.frame}

\item{cols}{A vector of column names to rid of NAs, defaults to all columns}

\item{value}{Defaults to 0, the value with which to replace NAs}
}
\description{
This function turns NA values to zero (or user specified value)
}
