% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_change.r
\name{pct_change}
\alias{pct_change}
\title{Percent Change Without Div by 0 Error}
\usage{
pct_change(prior, current, format = T, ...)
}
\arguments{
\item{prior}{A numeric object}

\item{current}{A numeric object}

\item{format}{If TRUE (default) then output is converted to a percentage with formattable, otherwise proportion}

\item{...}{Any arguments passed to formattable::percent, most likely digits}
}
\description{
Calculates the percent change of two numbers or vectors, if prior = 0 then percent change is 100%
}
\examples{
pct_change(0, 1) returns 100.00\%
pct_change(2, 3, format = F) returns 0.5

}
