% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_green_bars.R
\name{red_green_bars.}
\alias{red_green_bars.}
\alias{red_green_bars.g}
\alias{red_green_bars.r}
\title{Red/Green Color Bars}
\usage{
red_green_bars.g(x, data = NULL)

red_green_bars.r(x, data = NULL)
}
\arguments{
\item{x}{A numeric column of a data.frame or data.table}
}
\description{
These functions change the color bar of a cell of a kable obejct. The length of the color bar is determined
    for each tile by the absolute percentage of the value divided by the absolute max.
\itemize{
    \item\code{.g} makes positive values green, negative values red, and zero values black
    \item\code{.r} does the reverse, making positive values red
}
}
\note{
For best usage, apply the color transformation with mutate() and then pass the mutated object
    through kable() with escape = F and format = 'html'. If kableExtra::kable_styling() is used
    then use protect_latex = F.
}
