% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_green_text.R
\name{red_green_text.}
\alias{red_green_text.}
\alias{red_green_text.g}
\alias{red_green_text.r}
\title{Red/Green Font Formatters}
\usage{
red_green_text.g(x, data = NULL)

red_green_text.r(x, data = NULL)
}
\arguments{
\item{x}{A numeric column of a data.frame or data.table}
}
\description{
These functions change the font color of a cell of a kable obejct.
\itemize{
    \item\code{.g} makes positive values green, negative values red, and zero values black
    \item\code{.r} does the reverse, making positive values red
}
}
\note{
For best usage, apply the color transformation with mutate() and then pass the mutated object
    through kable() with escape = F and format = 'html'. If kableExtra::kable_styling() is used
    then use protect_latex = F.
}
