% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_green_tiles.R
\name{red_green_tiles.}
\alias{red_green_tiles.}
\alias{red_green_tiles.g}
\alias{red_green_tiles.r}
\alias{w2rg}
\title{Red/Green Tile Formatters}
\usage{
red_green_tiles.g(x, data = NULL)

red_green_tiles.r(x, data = NULL)

w2rg(x, green_up = T)
}
\arguments{
\item{x}{A numeric column of a data.frame or data.table}

\item{green_up}{Logical, default is TRUE where green values are positive. Though a user is unlikely to ever use w2rg by itself}
}
\description{
These functions shade the background of a kable cell either red or green, preserving the distance from zero which remains white.
\itemize{
    \item\code{.g} makes positive values green and negative values red
    \item\code{.r} makes negative values green and positive values red
    \item\code{w2rg()} is the function that creates the html colors to be passed to \code{red_green_tiles.}
   }
}
\note{
For best usage, apply the color transformation with mutate() and then pass the mutated object
    through kable() with escape = F and format = 'html'. If kableExtra::kable_styling() is used
    then use protect_latex = F.
}
