% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_setup.R
\name{rmarkdown_setup}
\alias{rmarkdown_setup}
\alias{setup2cols}
\alias{begin_left}
\alias{begin_right}
\alias{end_left}
\alias{end_right}
\alias{setup_pagewidth}
\alias{br}
\title{Rmarkdown Page and Column Setup, and Line Breaks!}
\usage{
setup2cols(left = 50, right = 50)

begin_left()

begin_right()

end_left()

end_right()

setup_pagewidth(width = 1700)

br(n = 1)
}
\arguments{
\item{left}{Integer, percent of the page width to devote to the left column, default is 50}

\item{right}{Integer, percent of the page width to devote to the right column, default is 50}

\item{width}{Integer, page width in px}

\item{n}{Integer, number of line breaks to include, deafult is 1}
}
\description{
Changing the page size and adding two column support for HTML RMarkdown documents is relatively easy
    but requires either altering an external .html file or a fair bit of typing in your .Rmd file. These functions
    make it easier to add these supports, fine-tuned to each .Rmd file.
    \itemize{
      \item\code{setup2cols()} sets the size of the left and right columns
      \item\code{setup_pagewidth()} sets the width of the page
      \item\code{begin_left()} and \code{begin_right} begin column support
      \item\code{end_left()} and \code{end_right} ends column support
      \item\code{br()} is thrown in for good measure as line breaks in Rmarkdown are finicky. Just call \code{br()}
         in an r chunk with the option \code{results='asis'}
         }
}
\details{
To use these functions put \code{setup2cols()} and/or \code{setup_pagewidth()} in a r chunk with the option
   \code{results='asis'} at the beginning of your document. Then, when you want column output, make another r chunk
   (also with \code{results='asis'}) and surround your desired output with \code{begin_left()} and \code{end_left()}.
   These functions were designed to be used with a for loop, hence why \code{results='hold'} is not used.
}
\note{
To use ggplot objects with \code{begin_left()} and \code{begin_right()} pass them through \code{plot()}.
    To use kable objects pass them through \code{htmltools::knit_print.html()}.
}
