% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_database.r
\name{send_to_database}
\alias{send_to_database}
\title{Save data.frame to Greenplum}
\usage{
send_to_database(data, name)
}
\arguments{
\item{data}{The data.frame to upload. Only supports numeric, character, and date column types.}

\item{name}{A character name to call the table in Greenplum.}
}
\description{
This function saves an R data.frame to a satabase using PostgreSQL's "INSERT INTO" function.
}
