% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dates.R
\name{set_dates}
\alias{set_dates}
\title{Create a data.table with 3,6,9,12 month periods}
\usage{
set_dates(last_month)
}
\arguments{
\item{last_month}{A character or Date object, formatted "\%Y-\%m-\%d". Ensure it is the first day of the final effective period.}
}
\description{
This function returns a 24x5 data.table; the first column is the 24 effective periods, the second to fifth
    columns assign each effper to either 'current' or 'prior' depending on the rolling window. The second column \code{col}
    is the 3 month rolling window and the others are identified by their number
}
\examples{
last_month <- as.Date('2019-02-01') or last_month <- '2019-02-01'
dates_df <- set_dates(last_month)
dates_df[!is.na(col), .(effper, col)] #could then be used to merge on to keep only certain effpers

}
