% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_line.r
\name{split_line}
\alias{split_line}
\title{Split a Character String Across 2 Lines}
\usage{
split_line(text)
}
\arguments{
\item{text}{A vector of character strings.}
}
\description{
This function splits a character string across two lines, keeping lines as even as possible.
    Replaces the middlemost " " (determined by string length) with "\\n"; and does not perform replacement if
    lines would have words with 3 or fewer characters.
}
\note{
To use with ggplot2 add \code{scale_x_discrete(labels = split_line)} to your ggobject.
}
\examples{
split_line(c('Hello World!', 'Goodbye Cruel World', 'To Myself'))
# Returns: 'Hello\\nWorld!' 'Goodbye\\nCruel World' 'To Myself'
}
