% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdate.r
\name{usdate}
\alias{usdate}
\title{Convert Date to USA style date}
\usage{
usdate(x, numeric = T)
}
\arguments{
\item{x}{The date object}

\item{numeric}{If TRUE, returns date in numbers (ie. 02/14/2019); if FALSE, returns Febraury 14, 2019. Default TRUE}
}
\description{
Converts a date object to month-day-year style USA date
}
