% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenplum_connect.r
\name{greenplum_connect}
\alias{greenplum_connect}
\alias{greenplum_disconnect}
\alias{create_secret}
\title{Setup Greenplum Database Connection}
\usage{
greenplum_connect(ask = T)

greenplum_disconnect()

create_secret()
}
\arguments{
\item{ask}{Logical, if \code{TRUE} (default) then RStudio will ask the user to input their database password,
if \code{FALSE} then password will be read from secret.rds}
}
\description{
This function establishes a RPostgreSQL database connection.

\itemize{
   \item\code{greenplum_connect()} creates a Greenplum connection object, \code{conn};
   \item\code{greenplum_disconnect()} disconnects \code{conn} and removes it from the global environment;
   \item\code{create_secret()} writes a password to 'C:/Users/<user>/Documents/Secrets/secret.rds',
        and creates a new directory if necessary

}
}
\note{
Database connection created with RPostgreSQL rather than RPostgres or RGreenplum as only RPostgreSQL can run multiple queries with
     one db call.
}
